Bug fix: CLI config generator no longer generates config with string floor number (fails validation)
Caused django to reject config when uploaded to node (endpoint validates received config), other than that didn't cause issues. No regression because I might remove int requirement (not sure why I added it in the first place).

+++ CLI/config_generator.py
@@ -228,7 +228,7 @@ class GenerateConfigFile:
 
         self.config['metadata'].update({
             'id': name,
-            'floor': floor,
+            'floor': int(floor),
             'location': location
         })
